pyproject.toml
[project]
name = "chainlit1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Saim-Hassan786", email = "saimaiengineer@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.4.1",
    "openai-agents>=0.0.4",
]

[project.scripts]
chainlit1 = "chainlit1:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
